+++ templates/item/parts/item-header.hbs
@@ -51,7 +51,7 @@
 						  style="border-top-right-radius: 12px;">
                         <label class="resource-label-m rollable" id="pdfLink">
 							{{localize 'FU.Source'}} </label>
-                        <input type="text" name="system.source.value" value="{{ system.source.value }}"
+                        <input type="text" name="system.source" value="{{ system.source }}"
 							   data-dtype="String" style="border-radius: 10px 0 0;" class="resource-inputs" />
                     </span>
                 </span>
